Prefer a base class for notify/registrations
